REF: implement arithmetic, comparison, logical ops on arrays (#28431)
